Moved curry and lift to separate file
