+++ src/Utility/Maybe.ts
@@ -99,20 +99,6 @@ export function maybe<A,B>(valueIfMaybeIsNothing:B, fnToApply:(value:A) => B,  m
     return maybeValue.map(fnToApply).orElse(valueIfMaybeIsNothing)
 }
 
-interface Applicative<Value> extends MonadDefinitions.Applicative<Value>{
-}
-
-export function liftA2<A,B,C>(fn:(arg1:A) => (arg2:B) => C, applicative1:Maybe<A>, applicative2:Maybe<B>):Maybe<C>
-export function liftA2<A,B,C>(fn:(arg1:A) => (arg2:B) => C, applicative1:Applicative<A>, applicative2:Applicative<B>):Applicative<C>{
-    return applicative1.map(fn).ap(applicative2)
-}
-
-export function curryA2<A,B,C>(fn:(arg1:A, arg2:B) => C): (arg1:A) => (arg2:B) => C{
-    return (arg1:A) => (arg2:B) => fn(arg1, arg2)
-}
-
-export function curryLiftA2<A,B,C>(fn:(arg1:A, arg2:B) => C, applicative1:Maybe<A>, applicative2:Maybe<B>):Maybe<C>
-export function curryLiftA2<A,B,C>(fn:(arg1:A, arg2:B) => C, applicative1:Applicative<A>, applicative2:Applicative<B>):Applicative<C>
-export function curryLiftA2<A,B,C>(fn:(arg1:A, arg2:B) => C, applicative1, applicative2){
-    return liftA2(curryA2(fn), applicative1, applicative2)
+export function orElseGet<A,B>(...fns: Array<(arg?:A) => Maybe<B>>): (arg?:A) => Maybe<B>{
+    return (arg?:A) => fns.reduce((result:Maybe<B>, fn:(arg:A) => Maybe<B>) => result.isSomething() ? result : fn(arg), Maybe.of(null))
 }
